backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
graphviz==0.20.3
inflection==0.5.1
packaging==24.1
pydotplus==2.0.2
PyJWT==2.8.0
pyparsing==3.1.2
pytz==2024.1
PyYAML==6.0.1
setuptools==70.1.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1